refactor: Category enum (member package -> post package)로 이동
